Animations oooh etc.

+++ code/app/src/main/java/com/example/booknet/Dialogs/ReviewCreateDialog.java
@@ -148,7 +148,7 @@ public class ReviewCreateDialog extends DialogFragment {
      */
     private void createReview() {
         //Get Info To Create Review
-        int score = (int) ratingInput.getRating();
+        float score = ratingInput.getRating();
         String message = commentField.getText().toString();
 
         //Create review
